feat: add GET loadAllWeather();
- 오픈 api에서 정보 받아와서 json으로 반환 성공

+++ src/main/java/efub/team4/backend_eweather/domain/weather/dto/OpenWeatherResponseDto.java
@@ -25,10 +25,10 @@ public class OpenWeatherResponseDto {
     private String fcstValue;
 
     /** 예보지점 X 좌표 */
-    private int nx;
+    private Long nx;
 
     /** 예보지점 Y 좌표 */
-    private int ny;
+    private Long ny;
 
     /**
      * 코드값 정보
@@ -63,7 +63,7 @@ public class OpenWeatherResponseDto {
     */
 
     @Builder
-    public OpenWeatherResponseDto(String baseDate, String baseTime, String fcstDate, String fcstTime, String category, String fcstValue, int nx, int ny) {
+    public OpenWeatherResponseDto(String baseDate, String baseTime, String fcstDate, String fcstTime, String category, String fcstValue, Long nx, Long ny) {
         this.baseDate = baseDate;
         this.baseTime = baseTime;
         this.fcstDate = fcstDate;
